Day21/Go/part1/part1_test.go
//	Copyright (c) Miloš Rackov 2022
//	Distributed under the Boost Software License, Version 1.0.
//	(See accompanying file LICENSE or copy at
//	https://www.boost.org/LICENSE_1_0.txt)

package part1

import (
	"os"
	"testing"
)

func TestPart1(t *testing.T) {
	testInput, err := os.ReadFile("../testinput.txt")
	if err != nil {
		t.Fatal(err)
	}

	type args struct {
		input string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "Provided input",
			args: args{
				input: string(testInput),
			},
			want: 152,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := Part1(tt.args.input); got != tt.want {
				t.Errorf("Part1() = %v, want %v", got, tt.want)
			}
		})
	}
}

func BenchmarkPart1(b *testing.B) {
	testInput, err := os.ReadFile("../input.txt")
	if err != nil {
		b.Fatal(err)
	}
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		Part1(string(testInput))
	}
}
